pyproject.toml
[tool.poetry]
name = "sepio-classifier"
version = "0.1.0"
description = ""
authors = ["Thiago S. Brasil <thiagosbrasil@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
firebase-admin = "^6.5.0"
google-cloud-firestore = "^2.16.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
